fix: Use a new Eulerian paths algorithm.
First we carefully convert each bidirectional path to a directional
one, using a heuristic that minimizing the number of Eulerian paths
and also the changes in angle.  For each bidirectional edge, we try to
keep a balance of input and output edges.  If there are multiple ways
to keep the balance then we find the solution that creates the least
change in direction.  Eventually, all bidirectional paths are turned
to directional ones.

Then we use the regular Hierholzer's algorithm on those directional
paths.